Integrator project

+++ my.planet.api/src/main/java/site/my/planet/controller/PersonalWorkController.java
@@ -3,6 +3,9 @@ package site.my.planet.controller;
 import java.util.ArrayList;
 import java.util.Optional;
 
+import javax.servlet.http.HttpServletRequest;
+
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -13,6 +16,8 @@ import lombok.SneakyThrows;
 import site.my.planet.service.PersonalWorkService;
 import site.my.planet.model.PersonalWork;
 import site.my.planet.model.PersonalWorkRequest;
+import site.my.planet.model.UserModel;
+import site.my.planet.security.JWTAutentication;
 
 @RestController
 @RequestMapping("/personal-work")
@@ -61,24 +66,54 @@ public class PersonalWorkController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable("id") long id) {
-        this.personalWorkService.delete(id);
+    public void delete(@PathVariable("id") long id,
+            HttpServletRequest request) {
+        if (verifyUserLogged(id, request)) {
+            this.personalWorkService.delete(id); 
+        }
     }
 
     @SneakyThrows
     @PutMapping("/{id}")
-    public void update(@PathVariable("id") long id, @RequestParam("personal-work") String personalWork,
+    public ResponseEntity<Object> update(@PathVariable("id") long id, @RequestParam("personal-work") String personalWork,
             @RequestParam("img") Optional<MultipartFile> multipartFile,
-            @RequestParam("deleteImage") Optional<String> deleteImage) {
+            @RequestParam("deleteImage") Optional<String> deleteImage,
+            HttpServletRequest request) {
+        if (verifyUserLogged(id, request)) {
+            PersonalWorkRequest personalWorkRequest = new ObjectMapper().readValue(
+                    personalWork, PersonalWorkRequest.class);
+    
+            return this.personalWorkService.update(id, personalWorkRequest, multipartFile, deleteImage); 
+        }
 
-        PersonalWorkRequest personalWorkRequest = new ObjectMapper().readValue(
-                personalWork, PersonalWorkRequest.class);
-
-        this.personalWorkService.update(id, personalWorkRequest, multipartFile, deleteImage);
+        return new ResponseEntity<>("Não autorizado", HttpStatus.UNAUTHORIZED);  
     }
 
     @PatchMapping("/public/visibility")
-    public PersonalWork updateVisibility(@RequestBody PersonalWork personalWork) {
-        return this.personalWorkService.updateVisibility(personalWork);
+    public ResponseEntity<Object> updateVisibility(@RequestBody PersonalWork personalWork,
+            HttpServletRequest request) {
+        if (verifyUserLogged(personalWork.getIdPersonalWork(), request)) {
+            return this.personalWorkService.updateVisibility(personalWork);
+        }
+
+        return new ResponseEntity<>("Não autorizado", HttpStatus.UNAUTHORIZED);  
+    }
+
+    public boolean verifyUserLogged(long id, HttpServletRequest request) {
+        String token = request.getHeader("Authorization");
+        Long subToken = Long.parseLong(new JWTAutentication().getSubToken(token));
+        Optional<PersonalWork> personalWork = this.personalWorkService.get(id);
+        
+        try {
+            if (subToken instanceof Long && subToken != null && personalWork.isPresent()) {
+                Long idUser = personalWork.get().getUser().getId();
+
+                return (idUser == subToken);
+            } 
+
+            return false;
+        } catch (Exception e) {
+            throw e;
+        }
     }
 }
